refactor: rename folders and components
