+++ class-components/src/components/resultsearch/ResultSearch.module.scss
