+++ class-components/src/components/resultsearch/ResultSearch.tsx
